training/requirements.txt
# Core ML
scikit-learn==1.4.1.post1
xgboost==2.0.3
numpy==1.26.4
pandas==2.2.1

# MLflow
mlflow>=2.14,<4.0

# Config
pyyaml==6.0.1

# Data validation
pandera==0.18.3

# HTTP (for dataset download)
requests==2.31.0

# Testing
pytest==8.1.1
pytest-cov==5.0.0

# Linting & formatting
black==24.3.0
ruff==0.3.4
mypy==1.9.0

# Type stubs
pandas-stubs==2.2.1.240316
